requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.in
#
click==6.7                # via python-dotenv
dj-database-url==0.4.2
django==1.11.1
djangorestframework==3.6.3
gunicorn==19.7.1
mysqlclient==1.3.10
psycopg2==2.7.1
py==1.4.33                # via pytest
pytest-django==3.1.2
pytest==3.0.7
python-dotenv==0.6.4
pytz==2017.2              # via django
